+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ## [5.3.5] - 2024-03-18
 - Fixed error with SlaveRep secs.behind being set to null.
 
+### Fixed
+- mysql_class.SlaveRep.upd_slv_time: Fixed down slave times to return "null" or "UNK".
+
 ### Changed
 - mysql_class.MasterRep.\_\_init\_\_: Added self.slaves attribute.
 - mysql_class.MasterRep.upd_mst_status: Set the self.slaves attribute.